backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
python-multipart
sentence-transformers
chromadb
pydantic
ollama